fix: skeleton location picker
