+++ app/src/main/kotlin/com/wire/android/ui/home/messagecomposer/AdditionalOptions.kt
@@ -29,6 +29,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import com.wire.android.ui.common.colorsScheme
 import com.wire.android.ui.home.conversations.model.UriAsset
+import com.wire.android.ui.home.messagecomposer.location.LocationPickerComponent
 import com.wire.android.ui.home.messagecomposer.recordaudio.RecordAudioComponent
 import com.wire.android.ui.home.messagecomposer.state.AdditionalOptionMenuState
 import com.wire.android.ui.home.messagecomposer.state.AdditionalOptionSelectItem
@@ -120,6 +121,12 @@ fun AdditionalOptionSubMenu(
             AdditionalOptionSubMenuState.AttachImage -> {}
             AdditionalOptionSubMenuState.Emoji -> {}
             AdditionalOptionSubMenuState.Gif -> {}
+            AdditionalOptionSubMenuState.Location -> {
+                LocationPickerComponent(
+                    onPickedLocation = {},
+                    onCloseLocation = {}
+                )
+            }
         }
     }
 }
