+++ app/src/main/kotlin/com/wire/android/ui/home/messagecomposer/state/AdditionalOptionMenuState.kt
@@ -34,6 +34,7 @@ enum class AdditionalOptionSubMenuState {
     AttachFile,
     AttachImage,
     Emoji,
+    Location,
     Gif;
 }
